+++ app/controllers/api/user_wants_controller.rb
@@ -2,27 +2,39 @@ module Api
   class UserWantsController < ApplicationController
 
     def create
-      @want = UserWant.new({
-        user_id: params[:user_want][:user_id],
-        record_id: params[:user_want][:record_id]
-      })
-
-      if @want.save
-        render :show
+      if uw_params[:user_id] == current_user.id
+        @want = UserWant.new(uw_params)
+        if @want.save
+          render :show
+        else
+          render json: @want.errors.full_messages, status: 422
+        end
       else
-        render json: @want.errors.full_messages, status: 422
+        head :forbidden
       end
     end
 
     def destroy
       want = UserWant.find(params[:id])
-      want.destroy!
 
-      render json: {}
+      if want.user_id == current_user.id
+        want.destroy!
+        render json: {}
+      else
+        head :forbidden
+      end
     end
 
     def show
       @want = UserWant.find(params[:id])
+    end
+
+    private
+
+    def uw_params
+      params
+        .require(:user_want)
+        .permit(:user_id, :record_id)
     end
   end
 end
